kalosm_language

Module context

Context for language models to consume.

Structs

Attributes

Document

A document is a piece of text with a title.

DocumentFolder

A folder full of documents.

DocumentFolderNotDirectoryError

The path to a document folder was not a directory

DocxDocument

A docx document that can be read from the file system.

HtmlDocument

An html document that can be read from the file system.

KnowledgeGraph

MdDocument

A markdown document that can be read from the file system.

Organic

PdfDocument

A pdf document that can be read from the file system.

PeopleAlsoAsk

RelatedSearches

RssFeed

A RSS feed that can be used to add documents to a search index.

SearchQuery

A search query that can be used to search for documents on the web.

SearchResult

Sitelinks

TextDocument

A text document that can be read from the file system.

Url

A parsed URL record.

Enums

ExtractDocumentError

An error that can occur when extracting a document from a URL.

FsDocument

A document that can be read from the file system.

FsDocumentError

An error that can occur when reading a document from the file system.

Lang

Represents a language following ISO 639-3 standard.

RssFeedError

An error that can occur when interacting with an RSS feed.

TextFileDecodeError

An error that can occur when decoding a text file.

Traits

IntoDocument

A trait for types that can be converted into a document.

IntoDocuments

A document that can be added to a search index.

Functions

search