kalosm_language::context

Trait IntoDocument

pub trait IntoDocument {
    type Error: Send + Sync + 'static;

    // Required method
    fn into_document(
        self,
    ) -> impl Future<Output = Result<Document, Self::Error>> + Send;
}

A trait for types that can be converted into a document.

Required Associated Types

type Error: Send + Sync + 'static

The error type that can occur when converting the type into a Document.

Required Methods

fn into_document( self, ) -> impl Future<Output = Result<Document, Self::Error>> + Send

Convert the type into a document.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.