kalosm_language

Module prelude

A prelude of commonly used items in kalosm-language

Re-exports

pub use futures_util::StreamExt as _;

pub use crate::context::*;

pub use crate::search::*;

pub use crate::vector_db::*;

Modules

kalosm_sample

kalosm-sample

Macros

bail

Bail out with the given error.

Structs

AnyOfSchema

A schema that matches any of the composite schemas

ArcParser

A boxed parser.

ArraySchema

A schema for an array

Bert

A bert embedding model. The main interface for this model is EmbedderExt.

BertBuilder

A builder for a Bert model

BertSource

A the source of a crate::Bert model

BooleanSchema

A schema for a boolean

BoxedChatConstraintsForType

A constraints for CreateDefaultChatConstraintsForType that work with boxed StructuredChatModels.

BoxedChatModel

A boxed ChatModel.

BoxedChatSession

A boxed ChatSession.

BoxedCompletionConstraintsForType

A constraints for CreateDefaultCompletionConstraintsForType that work with boxed TextCompletionModels.

BoxedStructuredChatModel

A boxed StructuredChatModel.

BoxedStructuredTextCompletionModel

A boxed BoxedStructuredTextCompletionModel.

BoxedTextCompletionModel

A boxed TextCompletionModel.

BoxedTextCompletionSession

A boxed TextCompletionSession.

CachedEmbeddingModel

Embedding models can be expensive to run. This struct wraps an embedding model with a cache that stores embeddings that have been computed before.

ChannelTextStream

A stream of text from a tokio channel.

Chat

Chat is a chat interface that builds on top of crate::ChatModel and crate::StructuredChatModel. It makes it easy to create a chat session with streaming responses, and constraints. Let’s start with a simple chat application:

ChatMessage

A single item in the chat history.

ChatResponseBuilder

A builder for a chat response. This is returned by Chat::add_message and can be modified until you start awaiting the response.

ChoiceParser

A parser for a choice of two parsers.

ChoiceParserState

State of a choice parser.

ConstSchema

A schema for a constant

DynEmbedder

A trait object for an embedder.

Embedding

Embeddings

EmbeddingInput

The input to an embedding model. This includes the text to be embedded and the type of embedding to output.

EmptyNumber

An error that can occur while parsing a float literal when trying to parse a number with no characters.

EnumSchema

A schema for an enum

FloatParser

A parser for a float.

FloatParserState

The state of an integer parser.

GenerationParameters

Parameters to use when generating text.

Html

An HTML tree.

I8Parser

A parser for i8.

I16Parser

A parser for i16.

I32Parser

A parser for i32.

I64Parser

A parser for i64.

IfThenSchema

A schema for an conditional schema

IndexParser

A parser that parses a sequence of parsers and returns the index of the first parser that succeeds

IndexParserState

The state of the IndexParser parser

IntegerParser

A parser for an integer.

IntegerParserState

The state of an integer parser.

IntegerSchema

A schema for an integer

InvalidDecimalLocation

An error that can occur while parsing a float literal when the number contains a decimal point in the wrong place.

InvalidSignLocation

An error that can occur while parsing a float literal when the number contains a sign in the wrong place.

JsonObjectSchema

A schema for an object

JsonPropertySchema

A schema for a property of an object

LazyParser

A parser that is lazily initialized.

LeadingZeroError

An error that can occur while parsing a float literal when the number starts with a leading zero.

LiteralMismatchError

The error type for a literal parser.

LiteralParser

A parser for a literal.

LiteralParserOffset

The state of a literal parser.

Llama

A quantized Llama language model with support for streaming generation.

LlamaBuilder

A builder with configuration for a Llama model.

LlamaSession

A Llama session with cached state for the current fed prompt

LlamaSource

A source for the Llama model.

MapOutputParser

A parser that maps the output of another parser.

NoConstraints

A marker type that indicates that no constraints were supplied.

NumberSchema

A schema for a number (floating point or integer)

OneLine

One line of text with some non-whitespace characters

OneLineError

An error that can occur when parsing a OneLine

OneLineState

The state of the OneLine parser

OneOfSchema

A schema that matches one of the composite schemas

OutOfRangeError

An error that can occur while parsing a float literal when the number is out of range.

ParagraphStream

A stream that output paragraphs of text at a time.

ParserError

An error that occurred while parsing.

RegexParser

A parser that uses a regex pattern to parse input.

RegexParserState

The state of a regex parser.

RepeatParser

A parser for a repeat of two parsers.

RepeatParserState

State of a repeat parser.

SegmentedStream

A stream that output segments of text at a time.

Sentence

A single word.

SentenceParser

A parser for a word.

SentenceStream

A stream that output sentences of text at a time.

SeparatedParser

A parser for a repeat of two parsers.

SeparatedParserState

State of a repeat parser.

SequenceParser

A parser for a sequence of two parsers.

StopOn

A parser that parses until a literal is found.

StopOnOffset

The state of a stop on literal parser.

StopOnParseError

An error that can occur while parsing a string literal.

StringParseError

An error that can occur while parsing a string literal.

StringParser

A parser for an ascii string.

StringParserState

The state of a literal parser.

StringSchema

A schema for a string

Task

A task session lets you efficiently run a task with a model. The task session will reuse the model’s cache to avoid re-feeding the task prompt repeatedly.

TextCompletionBuilder

A builder for a text completion response. This is returned by TextCompletionModelExt::complete and can be modified with TextCompletionBuilder::with_sampler and TextCompletionBuilder::with_constraints until you start awaiting the response.

ThenLazy

A parser that is initialized lazily based on the state of the previous parser.

U8Parser

A parser for u8.

U16Parser

A parser for u16.

U32Parser

A parser for u32.

U64Parser

A parser for u64.

WithInitialState

A parser with lazy initial state

Word

A single word.

WordParser

A parser for a word.

WordStream

A stream that output words of text at a time.

Enums

Either

A value that can be one of two types.

EmbeddingVariant

The type of embedding the model should output. For models that output different embeddings for queries and documents, this

MessageType

The type of a chat message

OwnedParseResult

A parser for a choice between two parsers.

ParseStatus

The state of a parser.

SchemaLiteral

A literal value in a schema

SchemaType

The type of a schema

SeparatedItemState

The state of the item in the separated parser.

SequenceParserState

State of a sequence parser.

StructureParser

A validator for a string

StructureParserState

The state of a structure parser.

ThenLazyParserState

State of a then lazy parser.

Traits

ChatModel

A trait for unstructured chat models. This trait is required for any chat models even if they do not support structured generation. While this trait is implemented for all chat models, most methods to use models that implement this trait are implemented in the ChatModelExt trait.

ChatModelExt

An extension trait for chat models with helpers for handling chat sessions. This trait is implemented automatically for all crate::ChatModels.

ChatSession

Chat Session

CreateChatSession

A trait for creating a chat session. While it the core trait every chat session implementation implements, most methods to use models that implement this trait are implemented in the ChatModelExt trait.

CreateDefaultChatConstraintsForType

A trait that defines the default constraints for a type with this chat model.

CreateDefaultCompletionConstraintsForType

A trait that defines the default constraints for a type with this model.

CreateParserState

A trait for a parser with a default state.

CreateTextCompletionSession

A trait for creating a text completion session for a model. While it the core trait every text completion model must implement, most methods to use models that implement this trait are implemented in the TextCompletionModelExt trait.

Embedder

A model that can be used to embed text. This trait is generic over the vector space that the model uses to help keep track of what embeddings came from which model.

EmbedderCacheExt

An extension trait for Embedder that allows for caching embeddings.

EmbedderExt

An extension trait for Embedder with helper methods for iterators, and types that can be converted into a string.

IntoChatMessage

A trait for types that can be converted into a chat message.

IntoEmbedding

Convert a type into an embedding with an embedding model.

ModelBuilder

A builder that can create a model asynchronously.

ModelConstraints

A type that can constrain the output of a model into a specific output type.

Parse

Data that can be parsed incrementally.

Parser

An incremental parser for a structured input.

ParserExt

An extension trait for parsers.

Pattern

A pattern that matches a character.

Schema

A description of the format of a type

SendCreateParserState

An auto trait for a Send parser with a default state.

StreamExt

An extension trait for Streams that provides a variety of convenient combinator functions.

StructuredChatModel

A trait for unstructured chat models that support structured generation. While this trait is implemented for all structured chat models, most methods to use models that implement this trait are implemented in the ChatModelExt trait.

StructuredTextCompletionModel

A trait for text completion models that support structured generation. While this trait is implemented for all structured text completion models, most methods to use models that implement this trait are implemented in the TextCompletionModelExt trait.

TextCompletionModel

A trait for unstructured text completion models. This trait is required for any text completion models even if they do not support structured generation. While this trait is implemented for all text completion models, most methods to use models that implement this trait are implemented in the TextCompletionModelExt trait.

TextCompletionModelExt

Text Completion Models

TextCompletionSession

Text Completion Session

TextStream

A stream of text. This is automatically implemented for all streams of something that acts like a string (String, &str).

Functions

prompt_input

A simple helper function for prompting the user for input.

Type Aliases

ParseResult

A result type for parsers.

Derive Macros

Parse

Derive a default JSON parser for a unit value, struct or enum.

Schema