kalosm_language::prelude::kalosm_sample

Trait ParserExt

pub trait ParserExt: Parser {
    // Provided methods
    fn otherwise<V>(self, other: V) -> ChoiceParser<Self, V>
       where V: Parser,
             Self: Sized { ... }
    fn or<V>(
        self,
        other: V,
    ) -> MapOutputParser<ChoiceParser<Self, V>, Self::Output>
       where V: Parser<Output = Self::Output>,
             Self: Sized { ... }
    fn then<V>(self, other: V) -> SequenceParser<Self, V>
       where V: Parser,
             Self: Sized { ... }
    fn then_lazy<V, F>(self, other: F) -> ThenLazy<Self, F>
       where Self: Sized,
             V: CreateParserState,
             F: Fn(&Self::Output) -> V { ... }
    fn ignore_output_then<V>(
        self,
        other: V,
    ) -> MapOutputParser<SequenceParser<Self, V>, <V as Parser>::Output>
       where V: CreateParserState,
             Self: Sized { ... }
    fn then_ignore_output<V>(
        self,
        other: V,
    ) -> MapOutputParser<SequenceParser<Self, V>, Self::Output>
       where V: CreateParserState,
             Self: Sized { ... }
    fn then_literal(
        self,
        literal: impl Into<Cow<'static, str>>,
    ) -> MapOutputParser<SequenceParser<Self, LiteralParser>, Self::Output>
       where Self: Sized { ... }
    fn repeat(self, length_range: RangeInclusive<usize>) -> RepeatParser<Self>
       where Self: Sized { ... }
    fn map_output<F, O>(self, f: F) -> MapOutputParser<Self, O, F>
       where Self: Sized,
             F: Fn(Self::Output) -> O { ... }
    fn boxed(self) -> ArcParser<Self::Output>
       where Self: Sized + CreateParserState + Send + Sync + 'static,
             Self::Output: Send + Sync + 'static,
             Self::PartialState: Send + Sync + 'static { ... }
    fn with_initial_state<F>(
        self,
        initial_state: F,
    ) -> WithInitialState<Self, F>
       where F: Fn() -> Self::PartialState + Clone,
             Self: Sized { ... }
}

An extension trait for parsers.

Provided Methods

fn otherwise<V>(self, other: V) -> ChoiceParser<Self, V>
where V: Parser, Self: Sized,

Parse this parser, or another other parser.

fn or<V>(self, other: V) -> MapOutputParser<ChoiceParser<Self, V>, Self::Output>
where V: Parser<Output = Self::Output>, Self: Sized,

Parse this parser, or another other parser with the same type

fn then<V>(self, other: V) -> SequenceParser<Self, V>
where V: Parser, Self: Sized,

Parse this parser, then the other parser.

fn then_lazy<V, F>(self, other: F) -> ThenLazy<Self, F>
where Self: Sized, V: CreateParserState, F: Fn(&Self::Output) -> V,

Parse this parser, then the other parser that is created base on the output of this parser.

fn ignore_output_then<V>( self, other: V, ) -> MapOutputParser<SequenceParser<Self, V>, <V as Parser>::Output>
where V: CreateParserState, Self: Sized,

Parse this parser, then the other parser while ignoring the current parser’s output.

fn then_ignore_output<V>( self, other: V, ) -> MapOutputParser<SequenceParser<Self, V>, Self::Output>
where V: CreateParserState, Self: Sized,

Parse this parser, then the other parser while ignoring the output of the other parser.

fn then_literal( self, literal: impl Into<Cow<'static, str>>, ) -> MapOutputParser<SequenceParser<Self, LiteralParser>, Self::Output>
where Self: Sized,

Parse this parser, then a literal. This is equivalent to .then_ignore_output(LiteralParser::new(literal)).

fn repeat(self, length_range: RangeInclusive<usize>) -> RepeatParser<Self>
where Self: Sized,

Repeat this parser a number of times.

fn map_output<F, O>(self, f: F) -> MapOutputParser<Self, O, F>
where Self: Sized, F: Fn(Self::Output) -> O,

Map the output of this parser.

fn boxed(self) -> ArcParser<Self::Output>
where Self: Sized + CreateParserState + Send + Sync + 'static, Self::Output: Send + Sync + 'static, Self::PartialState: Send + Sync + 'static,

Get a boxed version of this parser.

fn with_initial_state<F>(self, initial_state: F) -> WithInitialState<Self, F>
where F: Fn() -> Self::PartialState + Clone, Self: Sized,

Create a new parser with a different initial state