kalosm_language::prelude

Trait IntoEmbedding

pub trait IntoEmbedding {
    // Required methods
    fn into_embedding<E>(
        self,
        embedder: &E,
    ) -> impl Future<Output = Result<Embedding, <E as Embedder>::Error>> + Send
       where E: Embedder;
    fn into_query_embedding<E>(
        self,
        embedder: &E,
    ) -> impl Future<Output = Result<Embedding, <E as Embedder>::Error>> + Send
       where E: Embedder;
}

Convert a type into an embedding with an embedding model.

Required Methods

fn into_embedding<E>( self, embedder: &E, ) -> impl Future<Output = Result<Embedding, <E as Embedder>::Error>> + Send
where E: Embedder,

Convert the type into an embedding with the given embedding model.

fn into_query_embedding<E>( self, embedder: &E, ) -> impl Future<Output = Result<Embedding, <E as Embedder>::Error>> + Send
where E: Embedder,

Convert the type into a query embedding with the given embedding model.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.