kalosm_language

Module search

The index module contains different types of search indexes that can be used to search for crate::context::Documents created from crate::context::IntoDocument or crate::context::IntoDocuments

Structs

Chunk

A document snippet that can be used to display a snippet of a document.

DefaultSentenceChunker

The default sentence chunker. Unlike SentenceChunker, this is Send + Sync

HtmlSimplifier

Simplifies HTML by removing attributes and elements that does not affect the content

Hypothetical

Generates questions for a document.

HypotheticalBuilder

A builder for a hypothetical chunker.

SemanticChunker

A chunker that tries to create chunks of wroughly the same size while grouping together chunks with a similar meaning.

SentenceChunker

A Chunker that splits a string into sentences with a given SRX rules.

Summarizer

Generates summaries for a document.

Enums

ChunkStrategy

A strategy for chunking a document into smaller pieces.

HypotheticalChunkerError

An error that can occur when chunking a document with [HypotheticalChunker].

SummaryChunkerError

An error that can occur when chunking a document with [SummaryChunker].

Traits

Chunker

A strategy for chunking a document into smaller pieces.