kalosm_language::search

Trait Chunker

pub trait Chunker {
    type Error<E: Send + Sync + 'static>;

    // Required method
    fn chunk<E: Embedder + Send>(
        &self,
        document: &Document,
        embedder: &E,
    ) -> impl Future<Output = Result<Vec<Chunk>, Self::Error<E::Error>>> + Send;

    // Provided method
    fn chunk_batch<'a, I, E: Embedder + Send>(
        &self,
        documents: I,
        embedder: &E,
    ) -> impl Future<Output = Result<Vec<Vec<Chunk>>, Self::Error<E::Error>>> + Send
       where I: IntoIterator<Item = &'a Document> + Send,
             I::IntoIter: Send,
             Self: Sync { ... }
}

A strategy for chunking a document into smaller pieces.

Required Associated Types

type Error<E: Send + Sync + 'static>

The error type that can occur when chunking a document.

Required Methods

fn chunk<E: Embedder + Send>( &self, document: &Document, embedder: &E, ) -> impl Future<Output = Result<Vec<Chunk>, Self::Error<E::Error>>> + Send

Chunk a document into embedded snippets.

Provided Methods

fn chunk_batch<'a, I, E: Embedder + Send>( &self, documents: I, embedder: &E, ) -> impl Future<Output = Result<Vec<Vec<Chunk>>, Self::Error<E::Error>>> + Send
where I: IntoIterator<Item = &'a Document> + Send, I::IntoIter: Send, Self: Sync,

Chunk a batch of documents into embedded snippets.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.